+++ tskswp-rest-api/src/main/kotlin/com/tasksweeper/entities/ConsumableStatusEntity.kt
@@ -5,7 +5,7 @@ import org.jetbrains.exposed.sql.Table
 object ConsumableStatus : Table("tskswp.consumable_status") {
     val consumableName = varchar("consumable_name", 256).references(Consumable.name)
     val statusName = varchar("status_name", 256).references(Status.name)
-    val value = integer("value")
+    val value = long("value")
     val percentage = bool("percentage")
     val instant = bool("instant")
     override val primaryKey = PrimaryKey(consumableName, statusName, name = "consumable_status_pkey")
@@ -14,7 +14,7 @@ object ConsumableStatus : Table("tskswp.consumable_status") {
 data class ConsumableStatusDTO(
     val consumableName: String,
     val statusName: String,
-    val value: Int,
+    val value: Long,
     val percentage: Boolean,
     val instant: Boolean
 )
