+++ tskswp-rest-api/src/main/kotlin/com/tasksweeper/entities/EquipmentStatusEntity.kt
@@ -5,7 +5,7 @@ import org.jetbrains.exposed.sql.Table
 object EquipmentStatus : Table("tskswp.equipment_status") {
     val equipmentName = varchar("equipment_name", 256).references(Equipment.name)
     val statusName = varchar("status_name", 256).references(Status.name)
-    val value = integer("value")
+    val value = long("value")
     val percentage = bool("percentage")
     val instant = bool("instant")
     override val primaryKey = PrimaryKey(equipmentName, statusName, name = "equipment_status_pkey")
@@ -14,7 +14,7 @@ object EquipmentStatus : Table("tskswp.equipment_status") {
 data class EquipmentStatusDTO(
     val equipmentName: String,
     val statusName: String,
-    val value: Int,
+    val value: Long,
     val percentage: Boolean,
     val instant: Boolean
 )
